Move create-app.ts
