+++ src/common/twCreate.ts
@@ -4,7 +4,7 @@ import prompts from "prompts";
 import path from "path";
 import chalk from "chalk";
 import { validateNpmName } from "../create/helpers/validate-pkg";
-import { createApp, DownloadError } from "../create/helpers/create-app";
+import { createApp, DownloadError } from "./create-app";
 import { getPkgManager } from "../create/helpers/get-pkg-manager";
 
 let projectPath: string = "";
